requirements.txt
matplotlib
numpy
opencv-python
openslide-python
pillow
pysimplegui
scikit-image